refactor(frontend): extract useInlineFieldEditing composable (W4)
First stateful WorkflowDetail slice, gated by the new mount smoke. Move the
name/task-group/description inline-edit state machines (editing flags/values,
savingField, task-group options + the 11 start/cancel/save functions) into
composables/useInlineFieldEditing.js. workflow/workflowTaskIds are injected;
loadWorkflowDetail/buildDolphinConfigParams passed as lazy forward refs.
Returned refs/functions keep the same names so template bindings are unchanged.

Verified: WorkflowDetail mount smoke + full suite (27 files / 116 passed);
npm run build, npm run lint (0 errors). Verbatim move; WorkflowDetail.vue
2631 -> 2523.
